Moved things in preparation for tests and snapshots
