+++ src/client/src/ui/spotTile/components/Tile/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Tile'
